utility methods are moved from the base class to specific ones
